+++ src/core/python/miscdefs.cpp
@@ -94,6 +94,8 @@ static void define_trace(py::module& m)
                  "Counts pixel values and returns a list of (rgba, count) pairs.");
     quantize.def("k_modes", &utils::quantize::k_modes, py::arg("image"), py::arg("k"),
                  "Returns the k colors that appear most frequently in image.");
+    quantize.def("k_means", &utils::quantize::k_means, py::arg("image"), py::arg("k"), py::arg("max_iterations") = 100, py::arg("match") = 1,
+                 "Returns the k colors that are at the center of clusters.");
 }
 
 void define_utils(py::module& m)